+++ modules/portal-osgi-web/portal-osgi-web-portlet-tracker/build.gradle
@@ -1,6 +1,6 @@
 dependencies {
 	compile group: "com.liferay", name: "com.liferay.osgi.util", version: "2.0.7"
-	compile group: "com.liferay", name: "com.liferay.portal.servlet.jsp.compiler", version: "1.0.5"
+	compile group: "com.liferay", name: "com.liferay.portal.osgi.web.servlet.jsp.compiler", version: "1.0.5"
 	compile group: "com.liferay.portal", name: "portal-impl", version: liferay.portalVersion
 	compile group: "com.liferay.portal", name: "portal-service", version: liferay.portalVersion
 	compile group: "javax.portlet", name: "portlet-api", version: "2.0"
