+++ modules/portal-osgi-web/portal-osgi-web-servlet-jsp-compiler/src/test/java/com/liferay/portal/osgi/web/servlet/jsp/compiler/internal/TldURIUtilTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.servlet.jsp.compiler.internal;
+package com.liferay.portal.osgi.web.servlet.jsp.compiler.internal;
 
 import java.io.IOException;
