+++ modules/portal-osgi-web/portal-osgi-web-wab-extender/build.gradle
@@ -2,7 +2,7 @@ dependencies {
 	compile files("lib-patch/org.eclipse.equinox.http.servlet.jar")
 	compile group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
 	compile group: "com.liferay", name: "com.liferay.osgi.util", version: "2.0.7"
-	compile group: "com.liferay", name: "com.liferay.portal.servlet.jsp.compiler", version: "1.0.5"
+	compile group: "com.liferay", name: "com.liferay.portal.osgi.web.servlet.jsp.compiler", version: "1.0.5"
 	compile group: "com.liferay", name: "org.apache.axis", version: "1.4.LIFERAY-PATCHED-1"
 	compile group: "commons-fileupload", name: "commons-fileupload", transitive: false, version: "1.2.1"
 	compile group: "commons-io", name: "commons-io", transitive: false, version: "1.4"
