+++ portal-impl/src/portal.properties
@@ -6545,7 +6545,7 @@
         com.liferay.portal.servlet,\
         com.liferay.portal.servlet.filters.aggregate,\
         com.liferay.portal.servlet.filters.dynamiccss,\
-        com.liferay.portal.servlet.jsp.compiler,\
+        com.liferay.portal.osgi.web.servlet.jsp.compiler,\
         com.liferay.portal.spring.context,\
         com.liferay.portal.theme,\
         com.liferay.portal.util,\
