+++ source-formatter.properties
@@ -183,7 +183,7 @@
     # java.lang.reflect.Proxy.
     #
     proxy.excludes=\
-        modules/portal-osgi-web/portal-osgi-web-servlet-jsp-compiler/src/main/java/com/liferay/portal/servlet/jsp/compiler/JspServlet.java,\
+        modules/portal-osgi-web/portal-osgi-web-servlet-jsp-compiler/src/main/java/com/liferay/portal/osgi/web/servlet/jsp/compiler/JspServlet.java,\
         modules/portal-template/portal-template-freemarker/src/main/java/com/liferay/portal/template/freemarker/FreeMarkerManager.java,\
         portal-impl/src/com/liferay/portal/server/DeepNamedValueScanner.java,\
         portal-service/src/com/liferay/portal/kernel/util/ProxyUtil.java
